refactor: make endpointconfig handling more concise, separate logic, and cache result for subsequent serving
