+++ api/common/enums.js
@@ -1,0 +1,11 @@
+/**
+ * @typedef {Object} CacheKeys
+ * @property {'config'} CONFIG - Key for the config cache.
+ * @property {'defaultConfig'} DEFAULT_CONFIG - Key for the default config cache.
+ */
+const CacheKeys = {
+  CONFIG: 'config',
+  DEFAULT_CONFIG: 'defaultConfig',
+};
+
+module.exports = { CacheKeys };
