+++ api/server/services/Endpoints/index.js
@@ -1,0 +1,9 @@
+const { config } = require('./EndpointService');
+const loadAsyncEndpoints = require('./loadAsyncEndpoints');
+const loadDefaultConfig = require('./loadDefaultConfig');
+
+module.exports = {
+  config,
+  loadAsyncEndpoints,
+  loadDefaultConfig,
+};
